Extract FreeMarker into a separate module
