+++ ninja-freemarker/src/main/java/ninja/freemarker/FreemarkerModule.java
@@ -1,0 +1,10 @@
+package ninja.freemarker;
+
+import com.google.inject.AbstractModule;
+
+public class FreemarkerModule extends AbstractModule {
+	@Override
+	protected void configure() {
+		bind(TemplateEngineFreemarker.class);
+	}
+}
