+++ ninja-freemarker/src/test/java/ninja/freemarker/TemplateEngineFreemarkerI18nMethodTest.java
@@ -14,24 +14,32 @@
  * limitations under the License.
  */
 
-package ninja.template;
+package ninja.freemarker;
 
 import ch.qos.logback.core.Appender;
+
 import com.google.common.base.Optional;
 import com.google.common.collect.Lists;
+
 import freemarker.template.SimpleScalar;
 import freemarker.template.TemplateException;
 import freemarker.template.TemplateModel;
 import freemarker.template.TemplateModelException;
+
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+
 import ninja.Context;
 import ninja.Result;
+import ninja.freemarker.TemplateEngineFreemarkerI18nMethod;
 import ninja.i18n.Messages;
+
 import org.hamcrest.CoreMatchers;
 import org.junit.Test;
+
 import static org.junit.Assert.*;
+
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.rules.ExpectedException;
