+++ ninja-freemarker/src/test/resources/conf/application.conf
@@ -1,0 +1,47 @@
+##############################################################################
+# The main properties file to configure your application
+#
+# Properties can be prefixed by "%".
+# "%"" matches a mode you can set as system property. For instance when you
+# are using maven you can do: mvn -Dmode=test jetty:run
+#
+# Using that the following property:
+# %test.myproperty=test property
+# would overwrite the property
+# myproperty=my property
+#
+# You can define as many environments as you want. Simply set them in the mode.
+##############################################################################
+application.cookie.prefix=NINJA
+
+application.languages=en,de,fr-FR
+
+testproperty=testproperty_without_prefix
+%test.testproperty=test_testproperty
+%dev.testproperty=dev_testproperty
+
+## for testing getBoolean
+booleanTestTrue=true
+booleanTestFalse=false
+booleanNotValid=THIS_IS_NOT_A_VALID_BOOLEAN
+
+utf8Test=this is utf8: \u00F6\u00E4\u00FC
+
+## for testing getInteger
+integerTest=123456789
+integerNotValid=THIS_IS_NOT_A_VALID_INTEGER
+
+## testing composition of values
+serverName=http://myserver.com
+serverPort=80
+fullServerName=${serverName}:${serverPort}
+
+## testing working of getStringArray
+getOneElementStringArray=one
+getMultipleElementStringArrayWithSpaces=one , me
+getMultipleElementStringArrayWithoutSpaces=one,me
+
+
+# make sure that the constants are okay in NinjaConstant.*
+application.session.transferred_over_https_only=false
+application.secret = secret
