+++ ninja-freemarker/src/test/resources/conf/messages.properties
@@ -1,0 +1,25 @@
+#
+# Copyright (C) 2012-2014 the original author or authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+language=defaultlanguage
+message_with_placeholder=that`s the default language: {0}
+message_with_placeholder_date=that''s a date: {0, date, medium}
+
+
+a_property_only_in_the_defaultLanguage=prop
+
+## just a test that the parsing of properties does not break when a delimiter like "," is there:
+a_propert_with_commas=prop1, prop2, prop3
