+++ ninja-freemarker/src/test/resources/conf/messages_de.properties
@@ -1,0 +1,19 @@
+#
+# Copyright (C) 2012-2014 the original author or authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+language=deutsch
+message_with_placeholder=das ist der platzhalter: {0}
+message_with_placeholder_date=das ist ein datum: {0, date, medium}
