+++ ninja-servlet-archetype-simple/src/main/resources/archetype-resources/pom.xml
@@ -154,6 +154,11 @@
             <artifactId>h2</artifactId>
             <version>1.4.180</version>
         </dependency>
+        <dependency>
+            <groupId>org.ninjaframework</groupId>
+            <artifactId>ninja-freemarker</artifactId>
+            <version>${ninja.version}</version>
+        </dependency>
         <!-- If you want to deploy to a war please -->
         <!-- comment ninja-standalone dependency and  -->
         <!-- uncomment the dependency for ninja-servlet -->
