+++ ninja-servlet-jpa-blog-archetype/src/main/resources/archetype-resources/pom.xml
@@ -169,6 +169,12 @@
             <version>1.3.172</version>
         </dependency>
 
+        <dependency>
+            <groupId>org.ninjaframework</groupId>
+            <artifactId>ninja-freemarker</artifactId>
+            <version>${ninja.version}</version>
+        </dependency>
+
         <!-- If you want to deploy to a war please -->
         <!-- comment ninja-standalone dependency and  -->
         <!-- uncomment the dependency for ninja-servlet -->
